--- events/api/http/doc.go
@@ -1,10 +1,0 @@
-//
-// Copyright (c) 2018
-// Mainflux
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-
-// Package http contains implementation of edgex-app service HTTP API.
-
-package http
