--- events/postgres/doc.go
@@ -1,10 +1,0 @@
-//
-// Copyright (c) 2018
-// Mainflux
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-
-// Package postgres contains repository implementations using PostgreSQL as
-// the underlying database.
-package postgres
